% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grib.R
\name{era5_read_hourly_stars}
\alias{era5_read_hourly_stars}
\title{Read hourly}
\usage{
era5_read_hourly_stars(
  aoi_name = name,
  years = 2021:2021,
  months = 5:8,
  days = 1:31,
  hours = 0:23,
  variables = c("surface_net_solar_radiation"),
  download_dir = "."
)
}
\arguments{
\item{aoi_name}{string}

\item{years}{numeric from 1979-2021}

\item{months}{numeric from 1-12}

\item{days}{numeric from 1-31}

\item{hours}{numeric from 0-23}

\item{variables}{'surface_net_solar_radiation','2m_temperature','total_precipitation','snow_depth_water_equivalent', ...}

\item{download_dir}{download directory}
}
\value{
A matrix of the infile
}
\description{
This function
}
